Finished saving detected movies. Starting to movie the UI to MVVM Pattern

+++ Common/SQLiteInitializer.cs
@@ -8,18 +8,21 @@ namespace Frost.Common {
     /// <summary>If the databse doesn't exist it creates it and seeds it with data.</summary>
     public class SQLiteInitializer<TContext> : IDatabaseInitializer<TContext> where TContext : DbContext {
         private readonly string _initSQL;
+        private readonly bool _dropCreate;
 
 
         /// <summary>Initializes a new instance of the <see cref="SQLiteInitializer{TContext}"/> class.</summary>
         /// <param name="initSQL">The initialization SQL DDL/SDL.</param>
-        public SQLiteInitializer(string initSQL) {
+        /// <param name="dropCreate"></param>
+        public SQLiteInitializer(string initSQL, bool dropCreate = false) {
             _initSQL = initSQL;
+            _dropCreate = dropCreate;
         }
 
         public void InitializeDatabase(TContext context) {
             string dbName = context.Database.Connection.ConnectionString.Split('=')[1];
             try {
-                if (File.Exists(dbName)) {
+                if (_dropCreate && File.Exists(dbName)) {
                     try {
                         File.Delete(dbName);
                     }
@@ -30,9 +33,13 @@ namespace Frost.Common {
 
                 if (!File.Exists(dbName)) {
                     SQLiteConnection.CreateFile(dbName);
+                    SQLiteCommand.Execute(_initSQL, SQLiteExecuteType.NonQuery, context.Database.Connection.ConnectionString, new object());
+                    return;
                 }
 
-                SQLiteCommand.Execute(_initSQL, SQLiteExecuteType.NonQuery, context.Database.Connection.ConnectionString, new object());
+                if (_dropCreate) {
+                    SQLiteCommand.Execute(_initSQL, SQLiteExecuteType.NonQuery, context.Database.Connection.ConnectionString, new object());
+                }
             }
             catch (Exception e) {
                 Console.Error.WriteLine(e.Message);
